si4735: track chip bootup mode (FM/AM)
Track the bootup mode of the chip (FM or AM/SW/LW) in the si4735 module.
This will allow a cleanup of the cmd module, which will no longer need
to track this info internally.

Rationale: the bootup mode is a chip property and has repercussions for
which functions may be called on the chip.