cmd: track the band that the command layer is using
Track the band set in the command layer with a new enumerate. This
enumerate will supersede the command mode.

The command band is more granular than the si4735's chip mode, which is
either FM or AM/SW/LW. It is necessary to track the band somewhere, and
since it is not a property of the chip, it makes sense to keep it in the
command layer.